chore: optimize pseudorandom in create-schema
