+++ src/common/cfn-nag.ts
@@ -94,7 +94,7 @@ export function ruleRolePolicyWithWildcardResources(pattern: string, name: strin
     rules_to_suppress: [
       {
         id: 'W12',
-        reason: `Policy is generated by ${name}, * ${serviceName} resources for read only access or IAM limition.`,
+        reason: `Policy is generated by ${name}, * ${serviceName} resources for read only access or IAM limitation.`,
       },
     ],
   };
@@ -146,7 +146,7 @@ class AddCfnNagForCdkPath implements IAspect {
               || this.typeList.includes(node.cfnResourceType)
             )
           ) {
-            (node as CfnResource).addMetadata('cfn_nag', {
+            node.addMetadata('cfn_nag', {
               rules_to_suppress: nagConfig.rules_to_suppress,
             });
           }
